Create pubsub ontop of generic baseconnector
