+++ multicast.go
@@ -1,186 +1,61 @@
 package pubsub
 
 import (
+	"context"
 	"errors"
 	"io"
 	"iter"
 	"log/slog"
-	"sync"
-	"time"
-
-	"github.com/antoniomika/syncmap"
 )
 
 type PubSubMulticast struct {
-	Logger   *slog.Logger
-	Channels *syncmap.Map[string, *Channel]
+	*BaseConnector
+	Logger *slog.Logger
 }
 
-func (b *PubSubMulticast) Cleanup() {
-	toRemove := []string{}
-	for _, channel := range b.GetChannels() {
-		count := 0
-
-		for range channel.GetClients() {
-			count++
-		}
-
-		if count == 0 {
-			channel.Cleanup()
-			toRemove = append(toRemove, channel.ID)
-		}
-	}
-
-	for _, channel := range toRemove {
-		b.Channels.Delete(channel)
-	}
-}
-
-func (b *PubSubMulticast) GetChannels() iter.Seq2[string, *Channel] {
-	return b.Channels.Range
-}
-
-func (b *PubSubMulticast) GetClients() iter.Seq2[string, *Client] {
+func (p *PubSubMulticast) getClients(direction ChannelDirection) iter.Seq2[string, *Client] {
 	return func(yield func(string, *Client) bool) {
-		for _, channel := range b.GetChannels() {
-			channel.Clients.Range(yield)
+		for clientID, client := range p.GetClients() {
+			if client.Direction == direction {
+				yield(clientID, client)
+			}
 		}
 	}
 }
 
-func (b *PubSubMulticast) Connect(client *Client, channels []*Channel) (error, error) {
-	for _, channel := range channels {
-		dataChannel := b.ensureChannel(channel)
-		dataChannel.Clients.Store(client.ID, client)
-		defer func() {
-			client.Cleanup()
-			dataChannel.Clients.Delete(client.ID)
-
-			count := 0
-			for _, cl := range dataChannel.GetClients() {
-				if cl.Direction == ChannelDirectionInput || cl.Direction == ChannelDirectionInputOutput {
-					count++
-				}
-			}
-
-			if count == 0 {
-				for _, cl := range dataChannel.GetClients() {
-					cl.Cleanup()
-				}
-			}
-
-			b.Cleanup()
-		}()
-		client.Channels.Store(dataChannel.ID, dataChannel)
-	}
-
-	var (
-		inputErr  error
-		outputErr error
-		wg        sync.WaitGroup
-	)
-
-	// Pub
-	if client.Direction == ChannelDirectionInput || client.Direction == ChannelDirectionInputOutput {
-		wg.Add(1)
-		go func() {
-			defer wg.Done()
-			for {
-				data := make([]byte, 32*1024)
-				n, err := client.ReadWriter.Read(data)
-				data = data[:n]
-
-				channelMessage := ChannelMessage{
-					Data:      data,
-					ClientID:  client.ID,
-					Direction: ChannelDirectionInput,
-				}
-
-				if client.BlockWrite {
-					for {
-						count := 0
-						for _, channel := range client.GetChannels() {
-							for _, chanClient := range channel.GetClients() {
-								if chanClient.Direction == ChannelDirectionOutput {
-									count++
-								}
-							}
-						}
-
-						if count > 0 {
-							break
-						}
-
-						select {
-						case <-time.After(1 * time.Millisecond):
-							continue
-						case <-client.Done:
-							break
-						}
-					}
-				}
-
-				var sendwg sync.WaitGroup
-
-				for _, channel := range client.GetChannels() {
-					sendwg.Add(1)
-					go func() {
-						defer sendwg.Done()
-						select {
-						case channel.Data <- channelMessage:
-						case <-client.Done:
-						case <-channel.Done:
-						}
-					}()
-				}
-
-				sendwg.Wait()
-
-				if err != nil {
-					if errors.Is(err, io.EOF) {
-						return
-					}
-					inputErr = err
-					return
-				}
-			}
-		}()
-	}
-
-	// Sub
-	if client.Direction == ChannelDirectionOutput || client.Direction == ChannelDirectionInputOutput {
-		wg.Add(1)
-		go func() {
-			defer wg.Done()
-		mainLoop:
-			for {
-				select {
-				case data, ok := <-client.Data:
-					_, err := client.ReadWriter.Write(data.Data)
-					if err != nil {
-						outputErr = err
-						break mainLoop
-					}
-
-					if !ok {
-						break mainLoop
-					}
-				case <-client.Done:
-					break mainLoop
-				}
-			}
-		}()
-	}
-
-	wg.Wait()
-
-	return inputErr, outputErr
+func (p *PubSubMulticast) GetPipes() iter.Seq2[string, *Client] {
+	return p.getClients(ChannelDirectionInputOutput)
 }
 
-func (b *PubSubMulticast) ensureChannel(channel *Channel) *Channel {
-	dataChannel, _ := b.Channels.LoadOrStore(channel.ID, channel)
-	dataChannel.Handle()
-	return dataChannel
+func (p *PubSubMulticast) GetPubs() iter.Seq2[string, *Client] {
+	return p.getClients(ChannelDirectionInput)
+}
+
+func (p *PubSubMulticast) GetSubs() iter.Seq2[string, *Client] {
+	return p.getClients(ChannelDirectionOutput)
+}
+
+func (p *PubSubMulticast) connect(ctx context.Context, ID string, rw io.ReadWriter, channels []*Channel, direction ChannelDirection, blockWrite bool, replay bool) (error, error) {
+	client := NewClient(ID, rw, direction, blockWrite, replay)
+
+	go func() {
+		<-ctx.Done()
+		client.Cleanup()
+	}()
+
+	return p.Connect(client, channels)
+}
+
+func (p *PubSubMulticast) Pipe(ctx context.Context, ID string, rw io.ReadWriter, channels []*Channel, replay bool) (error, error) {
+	return p.connect(ctx, ID, rw, channels, ChannelDirectionInputOutput, false, replay)
+}
+
+func (p *PubSubMulticast) Pub(ctx context.Context, ID string, rw io.ReadWriter, channels []*Channel) error {
+	return errors.Join(p.connect(ctx, ID, rw, channels, ChannelDirectionInput, true, false))
+}
+
+func (p *PubSubMulticast) Sub(ctx context.Context, ID string, rw io.ReadWriter, channels []*Channel) error {
+	return errors.Join(p.connect(ctx, ID, rw, channels, ChannelDirectionOutput, false, false))
 }
 
 var _ PubSub = &PubSubMulticast{}
